+++ backend/canisters/user/impl/src/regular_jobs.rs
@@ -1,3 +1,4 @@
+use crate::updates::c2c_migrate_events_to_stable_memory::migrate_events_to_stable_memory_impl;
 use crate::Data;
 use utils::env::Environment;
 use utils::regular_jobs::{RegularJob, RegularJobs};
@@ -11,11 +12,13 @@ pub(crate) fn build() -> RegularJobs<Data> {
         5 * MINUTE_IN_MS,
     );
     let retry_deleting_files = RegularJob::new("Retry deleting files", retry_deleting_files, MINUTE_IN_MS);
+    let migrate_chat_events_to_stable_memory = RegularJob::new("Migrate chat events", migrate_chat_events_to_stable_memory, 0);
 
     RegularJobs::new(vec![
         check_cycles_balance,
         aggregate_direct_chat_metrics,
         retry_deleting_files,
+        migrate_chat_events_to_stable_memory,
     ])
 }
 
@@ -29,4 +32,8 @@ fn aggregate_direct_chat_metrics(_: &dyn Environment, data: &mut Data) {
 
 fn retry_deleting_files(_: &dyn Environment, _: &mut Data) {
     storage_bucket_client::retry_failed();
+}
+
+fn migrate_chat_events_to_stable_memory(_: &dyn Environment, data: &mut Data) {
+    migrate_events_to_stable_memory_impl(data, true);
 }